refactor: consolidate imports from githubService and taskService in GitHubIntegrationDetail
